transition to float

+++ linear.h
@@ -26,7 +26,7 @@
 struct vector {
         int size;
         int inc;
-        double *values;
+        float *values;
         int ref;
 };
 
@@ -44,7 +44,7 @@ inline struct vector *lualinear_newvector (lua_State *L, int size) {
 	vector->ref = LUA_NOREF;
 	luaL_getmetatable(L, LUALINEAR_VECTOR_METATABLE);
 	lua_setmetatable(L, -2);
-	vector->values = calloc(size, sizeof(double));
+	vector->values = calloc(size, sizeof(float));
 	if (vector->values == NULL) {
 		luaL_error(L, "cannot allocate values");
 	}
@@ -55,7 +55,7 @@ inline struct vector *lualinear_newvector (lua_State *L, int size) {
  * Pushes an existing vector onto the stack.
  */
 inline struct vector *lualinear_wrapvector (lua_State *L, int size,
-		double *values) {
+		float *values) {
 	struct vector *vector;
 
 	assert(size >= 1);
@@ -77,7 +77,7 @@ struct matrix {
         int cols;
         int ld;
         CBLAS_ORDER order;
-        double *values;
+        float *values;
         int ref;
 };
 
@@ -98,7 +98,7 @@ inline struct matrix *lualinear_newmatrix (lua_State *L, int rows, int cols,
 	matrix->ref = LUA_NOREF;
 	luaL_getmetatable(L, LUALINEAR_MATRIX_METATABLE);
 	lua_setmetatable(L, -2);
-	matrix->values = calloc((size_t)rows * cols, sizeof(double));
+	matrix->values = calloc((size_t)rows * cols, sizeof(float));
 	if (matrix->values == NULL) {
 		luaL_error(L, "cannot allocate values");
 	}
@@ -109,7 +109,7 @@ inline struct matrix *lualinear_newmatrix (lua_State *L, int rows, int cols,
  * Pushes an existing matrix onto the stack-
  */
 inline struct matrix *lualinear_wrapmatrix (lua_State *L, int rows, int cols,
-		CBLAS_ORDER order, double *values) {
+		CBLAS_ORDER order, float *values) {
 	struct matrix *matrix;
 
 	assert(rows >= 1 && cols >= 1);
